+++ packages/api-kintone/src/custMemo/index.ts
@@ -1,2 +1,3 @@
 export * from './deleteMemoById';
 export * from './getMemosByCustGroupId';
+export * from './saveMemo';
